+++ chrome/CHANGES.md
@@ -4,6 +4,10 @@ A free tool that provides an extra 'Add card' button on Gmail UI to add current 
 
 CHANGE LOG:
 
+Version 2.7.2.14
+-----------------
+- Check for ASL or ASF div for Refresh icon.
+
 Version 2.7.2.13
 -----------------
 - Restrict log from 1000 to 100 lines.